optimized tables, certs and exams

+++ app/database/migrations/2014_03_02_194802_create_forum_topics_table.php
@@ -16,7 +16,7 @@ class CreateForumTopicsTable extends Migration {
 			$table->increments('id');
 			$table->string('title', 60);
 			$table->integer('user_id');
-			$table->integer('topics');
+			$table->smallInteger('topics');
 			$table->boolean('has_attachments')->default(0);
 			$table->timestamps();
 		});
